Move metadata.py into RimSort.utils
