+++ RimSort/sort/dependencies.py
@@ -1,7 +1,6 @@
 from loguru import logger
-from typing import Any
 
-from util.metadata import MetadataManager
+from RimSort.utils.metadata import MetadataManager
 
 
 def gen_deps_graph(
