+++ RimSort/views/sub_views/inactive_mods_panel.py
@@ -18,7 +18,7 @@ from RimSort.controllers.settings_controller import SettingsController
 from RimSort.models.mod_list import ModListWidget
 from RimSort.models.mod_list_item import ModListItemInner
 from RimSort.utils.constants import SEARCH_DATA_SOURCE_FILTER_INDEXES
-from util.metadata import MetadataManager
+from RimSort.utils.metadata import MetadataManager
 
 
 class InactiveModList:
